test for queries on polymorphic relationship traversals
test for queries on polymorphic relationship traversals
